feat(encryption): add new encryption feature
Adding encryption for saving the api key in the config file.

Added encryption and decryption function in one file lockandkey.go
